Test handling more data types in training routine.
Dataloaders in Fed-BioMed should always return a tuple (data, target). In the base case, the `data` and `target`
are torch Tensors. However, they could also be lists, tuples or dicts. While the use is responsible for handling
these data types correctly in the `training_step` routine, we must make sure that the training routine as a whole
runs correctly.