Revise `DataManager` and implement `DataLoaderTypes` enum.
These changes are not mandatory per se as part of the POC, but thought of as a coherent
improvement on the existing code as part of the overall and `TrainingPlan` API revision.

Object: improve the way data loader type is specified by training plans and used through
the generic `DataManager` class.

Changes:
* Implement a `DataLoaderTypes` enum that maps concrete data loaders with frameworks'
  name, and is declared under `fedbiomed.common.data`.
* Add `TrainingPlan.data_loader_type()`, defined through a private class attribute,
  to specify the required type of data loader, that can be passed to `DataManager`
  and used to type-check available data loaders in train/test routines' code.
* Rename `DataManager.load(tp_type)` argument to `loader_type`.
* Remove deprecated `TrainingPlans` enum.
* Update `DataManager` unit tests.