Fix retrieval of share_persistent_buffers
Using TrainingArgs in the TrainingPlan is really complicated.
For example, self._training_args does not contain all the
training arguments, as one would expect.
I followed the current convention, which is to store a protected
variable with the argument requeste by the researcher. In the
future we may think of something better.