Update environments, scripts and packages from develop
This hotfix pulls the latests build scripts, Dockerfiles,
conda environments, and gui packages from develop. This
is needed because our current VPN containers cannot build
or run due to outdated or misconfigured packages. Essentially,
we aim to bring v4.4 up to date with the changes introduced
in PR 802 (related to issue 448).