New feature: keep same testing dataset over several `Rounds`.
Saving testing dataset over rounds by saving testing indexes

Context:
Before this update, it was not possible to test a model with a real validation dataset,
since the "validation dataset" was always changing from one `Round` to another.
This commit targets at fixing this issue.

Details:

Testing and training indexes are saved in the `NodeStateManager`, hence
are reload at the begining of each `Round`.

Includes one new variable `shuffle_testing_dataset`, in order to specify
to the user whether she/he/they want(s) to reset the testing dataset for this `Round` or not.
By default, shuffling / reseting dataset is disabled.