Feature/1439 dataclasses and register dataset (#1448)
This branch is followed by branch `feature/1440` that includes cherry
picked modifications from @alitolga for compatibility. All the work
associated to the dataclasses for the tables is based on the code
originally proposed by @alitolga.

This code includes:
- General purpose access to tables inside json files (singleton on
`fedbiomed/common/db.py`)
- Tables with dataclasses for control of input and functions to access
the 3 tables independently (dataset, dlp, dlb). All the interactions
with TinyDB are to be executed with dictionaries.
- _registry_controllers.py that gives a way to go from a string (names
compatible with `DatasetTypes`) to a controller
- values to be stored in dataset's table are recovered from
instantiated_controller including path, shape, type, others if
necessary.
- DatasetManager includes the change of the calls (API) and would be
deleted along with _dataset_db_manager.py and _dlp_db_manager.py as
DatasetManager is already much simpler than before and `aliases` are
meant to be removed for simplicity.

Code still to be fully tested. Submitted to open the discussion in case
of further modifications.

---------

Signed-off-by: Marc Vesin <marc.vesin@inria.fr>
Co-authored-by: Marc Vesin <marc.vesin@inria.fr>
Co-authored-by: Jhonatan TORRES <jho@interne-059a.sophia.inria.fr>
Co-authored-by: Ali Tolga Dinçer <alitolga.dincer@gmail.com>
Co-authored-by: Sergen CANSIZ <sergen.cansiz@inria.fr>